pyproject.toml
[project]
name = "closing-bell"
version = "1.0.0"
description = "종가매매 스크리너 - 거래대금 상위 종목 중 기술적 분석 기반 TOP 3 선정"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.9"
keywords = [
    "stock",
    "screening",
    "trading",
    "korea-investment",
    "technical-analysis",
    "CCI",
    "discord",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "APScheduler>=3.10.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
dashboard = [
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "pandas>=2.0.0",
]
all = [
    "closing-bell[dev,dashboard]",
]

[project.urls]
Homepage = "https://github.com/yourusername/closing-bell"
Documentation = "https://github.com/yourusername/closing-bell#readme"
Repository = "https://github.com/yourusername/closing-bell"
Issues = "https://github.com/yourusername/closing-bell/issues"

[project.scripts]
closing-bell = "main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "dashboard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "build", "dist"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "tests/",
    ".venv/",
]
